Make create student modal, install modals and dropzone
