+++ client/package.json
@@ -13,7 +13,9 @@
     "18": "^0.0.0",
     "@mantine/carousel": "^7.16.1",
     "@mantine/core": "^7.15.2",
+    "@mantine/dropzone": "^7.16.3",
     "@mantine/hooks": "^7.15.2",
+    "@mantine/modals": "^7.16.3",
     "embla-carousel-react": "^7.1.0",
     "react": "^18.3.1",
     "react-dom": "^18.3.1",
