+++ client/src/main.jsx
@@ -7,11 +7,14 @@ import App from "./App.jsx";
 import "@mantine/core/styles.css";
 import { MantineProvider } from "@mantine/core";
 import '@mantine/carousel/styles.css';
-
+import { ModalsProvider } from '@mantine/modals';
+import '@mantine/dropzone/styles.css';
 createRoot(document.getElementById("root")).render(
   <StrictMode>
     <MantineProvider>
-      <App />
+      <ModalsProvider>
+        <App />
+      </ModalsProvider>
     </MantineProvider>
   </StrictMode>
 );
